Add a colorcolumn character option.
This option allows a character to be displayed in the colorcolumn. This
allows the colorcolumn ot have a traditional line if the colorcolumn
character is set to a vertical line '│' or even a series of dots for a
':'.

The option is a part of the 'listchars' setting under the name
'colorcol'. So 'set listchars=colorcol:│' will work.