Add user-registers for arbitrary registers.
This allows users to define behaviors for arbitrary registers. These
registers can be any character including multibyte characters. This
means that any character may be used as a register and if that register
is not a builtin register, it will defer to a user-defined vimscript
function for behavior.

This is done throw an option called 'userregfun'

The function that 'userregfun' defines is a function that takes 3
arguments:

   action - Either set to "put" or "yank"
   register - The character representing the register.
   content - If the action is "yank" this string contains the content
             yanked.

Multibyte registers are still broken for expressions. So while

  let @&=xyz

Works as expected,

  let @λ=xyz

will still throw a parse error.